Move RouterContract into Contracts namespace
